Add files via upload
This is my tag-along code for the Hamoye Stage E project. 

This project seeks to explore and build time series forecasting models for measurements of electricity prices of a District given the total amount of electricity consumed in that District from 2010 to 2021.

Data Set Information:

The archive contains 92,016 measurements of electricity data gathered between September 2010 and February 2021.